+++ testing/primitive.py
@@ -16,18 +16,18 @@ def test():
 try:
 	test()
 except Exception as ee:
-	print(repr(ee))
+	print("|A|\t", repr(ee))
 	stackTraceList = traceback.extract_stack(ee.__traceback__.tb_frame)
-	del stackTraceList[0]
+	#del stackTraceList[0]
 	for frame in stackTraceList:
-		print("\t", frame)
+		print("|B|\t\t", frame)
 
 	if ee.__context__:
-		print(repr(ee.__context__))
+		print("|C|\t", repr(ee.__context__))
 		stackTraceList = traceback.extract_stack(ee.__context__.__traceback__.tb_frame)
-		del stackTraceList[0]
+		#del stackTraceList[0]
 		for frame in stackTraceList:
-			print("\t", frame)
+			print("|D|\t\t", frame)